Split init_passes.h into init_{iree,mlir}_passes.h
