+++ bindings/python/pyiree/compiler/compiler.cc
@@ -27,7 +27,8 @@
 #include "iree/compiler/Dialect/VM/Transforms/Passes.h"
 #include "iree/tools/init_compiler_modules.h"
 #include "iree/tools/init_dialects.h"
-#include "iree/tools/init_passes.h"
+#include "iree/tools/init_iree_passes.h"
+#include "iree/tools/init_mlir_passes.h"
 #include "iree/tools/init_targets.h"
 #include "llvm/Support/CommandLine.h"
 #include "llvm/Support/PrettyStackTrace.h"
