+++ iree/tools/opt_main.cc
@@ -22,7 +22,8 @@
 #include "iree/compiler/Dialect/HAL/Conversion/Passes.h"
 #include "iree/tools/init_compiler_modules.h"
 #include "iree/tools/init_dialects.h"
-#include "iree/tools/init_passes.h"
+#include "iree/tools/init_iree_passes.h"
+#include "iree/tools/init_mlir_passes.h"
 #include "iree/tools/init_targets.h"
 #include "iree/tools/init_xla_dialects.h"
 #include "llvm/Support/CommandLine.h"
